[plugin] fix #4621: apply only valid configurations as schemas
